+++ R/rd-examples.R
@@ -3,6 +3,31 @@ roxy_tag_parse.roxy_tag_examples <- function(x) {
   tag_examples(x)
 }
 #' @export
+roxy_tag_parse.roxy_tag_examplesIf <- function(x) {
+  lines <- unlist(strsplit(x$raw, "\r?\n"))
+
+  condition <- lines[1]
+  tryCatch(
+    suppressWarnings(parse(text = condition)),
+    error = function(err) {
+      roxy_tag_warning(x, "failed to parse condition of @examplesIf")
+    }
+  )
+
+  dontshow <- paste0(
+    "\\dontshow{if (",
+    condition,
+    ") (if (getRversion() >= \"3.4\") withAutoprint else force)(\\{ # examplesIf}"
+  )
+
+  x$raw <- paste(
+    c(dontshow, lines[-1], "\\dontshow{\\}) # examplesIf}"),
+    collapse = "\n"
+  )
+
+  x <- tag_examples(x)
+}
+#' @export
 roxy_tag_parse.roxy_tag_example <- function(x) {
   x <- tag_value(x)
 
@@ -17,6 +42,10 @@ roxy_tag_parse.roxy_tag_example <- function(x) {
 
 #' @export
 roxy_tag_rd.roxy_tag_examples <- function(x, base_path, env) {
+  rd_section("examples", x$val)
+}
+#' @export
+roxy_tag_rd.roxy_tag_examplesIf <- function(x, base_path, env) {
   rd_section("examples", x$val)
 }
 #' @export