+++ R/rd-describe-in.R
@@ -23,32 +23,50 @@ topic_add_describe_in <- function(topic, block, env) {
   }
 
   dest <- find_object(tag$val$name, env)
-  label <- build_label(block$object, dest, block)
-  if (is.null(label))
-    return()
+  metadata <- build_minidesc_metadata(src = block$object, dest = dest)
 
-  topic$add(rd_section_minidesc(
-    df = subsection(label = label$label, desc = tag$val$description),
-    method = label$method,
-    generic = label$generic
-  ))
+  topic$add(rlang::exec(rd_section_minidesc, !!!c(
+    name = block$object$topic,
+    desc = tag$val$description,
+    metadata
+  )))
   dest$topic
 }
 
 # Field -------------------------------------------------------------------
 
 #' Record data for minidescription sections from `@describeIn`
-#' @param df a dataframe as created by [subsection()].
-#' @param method giving whether method generic a generic (S3 or S4)
-#' @param generic Name of the generic in the destination the method generic, otherwise empty string (`""`).
+#' 
+#' @param name name of the source function.
+#' @param desc description passed to `@describeIn`.
+#' @param extends how the source function extends the destination function:
+#' - `"generic"` if the source extends a (S3 or S4) generic in the destination,
+#' - `"class"` if the source extends an informal S3 or formal S4 constructor
+#'    in the destination.
+#'    For S3, there is always only *one* class.
+#'    For S4, the methods' signature is used instead, to cover multiple dispatch.
+#' - `""` (default) otherwise.
+#' @param generic,class name of the generic and class that is being extended by 
+#' the method, otherwise empty string (`""`).
+#' @return a dataframe with one row for each `@describeIn`, wrapped inside 
+#' `rd_section()`
 #' @noRd
-rd_section_minidesc <- function(df, method = FALSE, generic = "") {
-  stopifnot(is_scalar_logical(method))
+rd_section_minidesc <- function(name,
+                                desc,
+                                extends = c("", "generic", "class"),
+                                generic = "",
+                                class = "") {
+  stopifnot(is_scalar_character(name))
+  stopifnot(is_character(desc))
+  rlang::arg_match(extends)
   stopifnot(is_scalar_character(generic))
-  data <- cbind(
-    method = method,
+  stopifnot(is_scalar_character(class))
+  data <- data.frame(
+    name = name,
+    desc = desc,
+    extends = extends,
     generic = generic,
-    df, 
+    class = class,
     stringsAsFactors = FALSE
   )
   rd_section(
@@ -66,64 +84,63 @@ merge.rd_section_minidesc <- function(x, y, ...) {
   )
 }
 
+# Rd Output -------------------------------------------------------------------
+
 #' @export
 format.rd_section_minidesc <- function(x, ...) {
   section_title <- "Related Functions and Methods"
-
-  groups <- unique(x$value[, c("method", "generic")])
-  out <- purrr::map2_chr(groups$method, groups$generic, function(method, generic) {
-    if (method) {
-      subsection_title <- "Methods extending"
-      if (generic == "") {
-        subsection_title <- paste0(subsection_title, " other generics (by generic):")
-      } else {
-        subsection_title <- paste0(
-          subsection_title,
-          " \\code{",
-          escape(generic),
-          "} generic (by class):"
-        )
-      }
-    } else {
-      subsection_title <- "Functions"
-    }
-    df <- x$value
-    subsection_body <- format_subsection(
-      df[df$method == method & df$generic == generic, c("label", "desc")]
-    )
-    paste0(
-      "\\subsection{", subsection_title, "}{\n",
-      subsection_body,
-      "}\n"
-    )
-  })
+  
+  df <- x$value
+  subsection_by <- unique(df[, "extends"])
+  section_body <- purrr::map_chr(subsection_by, format_subsections, df)
+  
   paste0(
     "\\section{", section_title, "}{\n",
-    paste0(out, collapse = "\n"),
+    paste0(section_body, collapse = "\n"),
     "}\n"
   )
 }
 
-subsection <- function(label, desc) {
-  stopifnot(is_character(label), is_character(desc))
-  stopifnot(length(label) == length(desc))
-  data.frame(
-    stringsAsFactors = FALSE,
-    label = label,
-    desc = desc
+format_subsections <- function(extends, df) {
+  class <- unique(df[df$extends == extends, "class"])
+  generic <- unique(df[df$extends == extends, "generic"])
+  subsection_title <- switch(
+    extends,
+    class = paste0(
+      "Methods extending \\code{", escape(class), "} class (by generic):"
+    ),
+    generic = paste0(
+      "Methods extending \\code{", escape(generic), "} generic (by class):"
+    ),
+    "Functions"
+  )
+
+  list_by <- unique(df[df$extends == extends, , drop = FALSE])
+  subsection_body <- purrr::pmap_chr(list_by, format_subsection)
+  subsection_body <- paste0(subsection_body, collapse = "\n")
+
+  paste0(
+    "\\subsection{", subsection_title, "}{\n", subsection_body, "}",
+    collapse = "\n"
   )
 }
 
-format_subsection <- function(df) {
+format_subsection <- function(name, desc, extends, generic, class) {
+  label <- switch(
+    extends,
+    "class" = generic,
+    "generic" = class,
+    name
+  )
   paste0(
-    "\\itemize{\n",
+  "\\itemize{\n",
     paste0(
       "\\item \\code{",
-      escape(df$label),
-      "}: ", df$desc,
+      escape(label),
+      "}: ", desc,
       collapse = "\n"
     ),
-    "\n}\n"
+    "\n}"
   )
 }
 
@@ -142,59 +159,62 @@ find_object <- function(name, env) {
   }
 }
 
-build_label <- function(src, dest, block) {
+#' Build metadata for how to present `@describeIn` tag
+#' @return list of character scalars named `extends`, `generic`, `class`.
+#' See rd_section_minidesc() for details.
+#' @noRd
+build_minidesc_metadata <- function(src, dest) {
   src_type <- class(src)[1]
   dest_type <- class(dest)[1]
   dest_name <- as.character(dest$topic)
 
-  # default to fallback: function + topic
-  method <- FALSE
-  generic <- ""
-  label <- src$topic
-
+  # fallback defaults
+  extends <- class <- generic <- ""
+  
   if (src_type == "s3method") {
-    if (dest_type == "s3generic" && attr(src$value, "s3method")[1] == dest_name) {
-      # applies only to those methods which actually extend the generic in dest
-      method <- TRUE
-      generic <- dest_name
-      label <- attr(src$value, "s3method")[2]
-    } else if (has_constructor(dest_name, src)) {
-      method <- TRUE
-      generic <- ""
-      label <- attr(src$value, "s3method")[1]
+    generic <- attr(src$value, "s3method")[1]
+    class <- attr(src$value, "s3method")[2]
+    if (dest_type == "s3generic" && generic == dest_name) {
+      # src method fits dest generic
+      extends <- "generic"
+    } else if (fits_constructor(dest_name, src)) {
+      # src method fits informal dest constructor (heuristically)
+      extends <- "class"
     }
   } else if (src_type == "s4method") {
-    if (dest_type == "s4class") {
-      # Label S4 methods in class with their generic
-      method <- TRUE
-      generic <- ""
-      label <- src$value@generic
-    } else if (dest_type == "s4generic") {
-      # Label S4 methods in generic with their signature
-      method <- TRUE
-      generic <- dest_name
-      sig <- src$value@defined
-      if (length(sig) == 1) {
-        label <- sig
-      } else {
-        label <- paste0(names(sig), " = ", sig, collapse = ",")
-      }
+    generic <- as.character(src$value@generic)
+    class <- sig2class(src$value@defined)
+    if (dest_type == "s4generic") {
+      # TODO must test whether src method fits dest generic
+      extends <- "generic"
+    } else if (dest_type == "s4class") {
+      extends <- "class"
+      # TODO must test whether src method fits dest constructor
     }
   }
-  label <-  as.character(label)
+  list(extends = extends, generic = generic, class = class)
+}
 
-  list(method = method, generic = generic, label = label)
+#' Turn S4 signatures into a character scalar
+#' @noRd
+sig2class <- function(sig) {
+  if (length(sig) == 1) {
+    class <- sig
+  } else {
+    class <- paste0(names(sig), " = ", sig, collapse = ",")
+  }
+  as.character(class)
 }
 
 #' Tests if destination is a constructor for class of src
 #' 
 #' No formal test is possible, these are heuristics.
 #' @noRd
-has_constructor <- function(dest_name, src) {
+fits_constructor <- function(dest_name, src) {
   src_class <- attr(src$value, "s3method")[2]
   # assuming that dest is class constructor, when class == dest name
   if (src_class == dest_name) TRUE
-  else has_constructor_disambig_class(src_class, dest_name)
+  else fits_constructor_disambig_class(src_class, dest_name)
 }
 
 #' Test if class of src is a disambiguated version of dest
@@ -209,7 +229,7 @@ has_constructor <- function(dest_name, src) {
 #' This function checks for these kinds of heuristic matches between src_class
 #' and dest_name.
 #' @noRd
-has_constructor_disambig_class <- function(src_class, dest_name, pkg_name) {
+fits_constructor_disambig_class <- function(src_class, dest_name, pkg_name) {
   evalenv <- roxy_meta_get("env")
   # This should only happen in our test cases
   if (is.null(evalenv)) evalenv <- parent.frame()