+++ tests/testthat/test-rd-describe-in.R
@@ -1,4 +1,4 @@
-test_that("@describeIn generic captures s3 method class", {
+test_that("@describeIn generic destination captures s3 method source", {
   out <- roc_proc_text(rd_roclet(), "
     #' Title
     f <- function(x) UseMethod('f')
@@ -7,13 +7,12 @@ test_that("@describeIn generic captures s3 method class", {
     #'
     f.a <- function(x) 1
   ")[[1]]
-
-  expect_equal(out$get_value("minidesc")$method, TRUE)
+  expect_equal(out$get_value("minidesc")$extends, "generic")
   expect_equal(out$get_value("minidesc")$generic, "f")
-  expect_equal(out$get_value("minidesc")$label, "a")
+  expect_equal(out$get_value("minidesc")$class, "a")
 })
 
-test_that("@describeIn generic captures s4 method class", {
+test_that("@describeIn generic destination captures s4 method source", {
   out <- roc_proc_text(rd_roclet(), "
     #' Title
     setGeneric('f', function(x) standardGeneric('f'))
@@ -21,13 +20,14 @@ test_that("@describeIn generic captures s4 method class", {
     #' @describeIn f Method for a
     setMethod(f, signature('a'), function(x) 1)
   ")[[1]]
+  out$get_value("minidesc")
 
-  expect_equal(out$get_value("minidesc")$method, TRUE)
+  expect_equal(out$get_value("minidesc")$extends, "generic")
   expect_equal(out$get_value("minidesc")$generic, "f")
-  expect_equal(out$get_value("minidesc")$label, "a")
+  expect_equal(out$get_value("minidesc")$class, "a")
 })
 
-test_that("@describeIn class captures s3 generic name", {
+test_that("@describeIn constructor destination captures s3 method source", {
   out <- roc_proc_text(rd_roclet(), "
     #' Title
     boo <- function() structure(list(), class = 'boo')
@@ -37,12 +37,12 @@ test_that("@describeIn class captures s3 generic name", {
     mean.boo <- function(x) 1
     ")[[1]]
 
-  expect_equal(out$get_value("minidesc")$method, TRUE)
-  expect_equal(out$get_value("minidesc")$generic, "")
-  expect_equal(out$get_value("minidesc")$label, "mean")
+  expect_equal(out$get_value("minidesc")$extends, "class")
+  expect_equal(out$get_value("minidesc")$generic, "mean")
+  expect_equal(out$get_value("minidesc")$class, "boo")
 })
 
-test_that("@describeIn class captures s4 generic name", {
+test_that("@describeIn constructor destination captures s4 method source", {
   out <- roc_proc_text(rd_roclet(), "
     setGeneric('mean')
 
@@ -52,13 +52,13 @@ test_that("@describeIn class captures s4 generic name", {
     #' @describeIn a mean method
     setMethod('mean', 'a', function(x) 1)
     ")[[1]]
-
-  expect_equal(out$get_value("minidesc")$method, TRUE)
-  expect_equal(out$get_value("minidesc")$generic, "")
-  expect_equal(out$get_value("minidesc")$label, "mean")
+  out$get_value("minidesc")
+  expect_equal(out$get_value("minidesc")$extends, "class")
+  expect_equal(out$get_value("minidesc")$generic, "mean")
+  expect_equal(out$get_value("minidesc")$class, "a")
 })
 
-test_that("@describeIn class captures function name", {
+test_that("@describeIn function destination captures function source", {
   out <- roc_proc_text(rd_roclet(), "
     #' Title
     f <- function(x) 1
@@ -67,9 +67,10 @@ test_that("@describeIn class captures function name", {
     f2 <- function(x) 1
     ")[[1]]
 
-  expect_equal(out$get_value("minidesc")$method, FALSE)
+  expect_equal(out$get_value("minidesc")$name, "f2")
+  expect_equal(out$get_value("minidesc")$extends, "")
   expect_equal(out$get_value("minidesc")$generic, "")
-  expect_equal(out$get_value("minidesc")$label, "f2")
+  expect_equal(out$get_value("minidesc")$class, "")
 })
 
 test_that("@describeIn class captures function name with data", {
@@ -82,7 +83,7 @@ test_that("@describeIn class captures function name with data", {
     f2 <- function(x) 1
     ")[[1]]
 
-  expect_equal(out$get_value("minidesc")$label, "f2")
+  expect_equal(out$get_value("minidesc")$name, "f2")
 })
 
 test_that("@describeIn class captures function description", {
@@ -97,6 +98,47 @@ test_that("@describeIn class captures function description", {
   expect_equal(out$get_value("minidesc")$desc, "A")
 })
 
+test_that("Multiple @describeIn functions combined into one", {
+  example <- test_path("roxygen-block-describe-in-functions.R")
+  out <- roc_proc_text(rd_roclet(), brio::read_file(example))[[1]]
+
+  expect_equal(out$get_value("minidesc")$name, c("square", "cube"))
+  expect_equal(out$get_value("minidesc")$extends, c("", ""))
+  expect_equal(out$get_value("minidesc")$generic, c("", ""))
+  expect_equal(out$get_value("minidesc")$class, c("", ""))
+})
+
+test_that(
+  "Multiple @describeIn methods and others are combined into a generic", {
+    example <- test_path("roxygen-block-describe-in-method-in-gen.R")
+    out <- roc_proc_text(rd_roclet(), brio::read_file(example))[[1]]
+    expect_equal(
+      out$get_value("minidesc")$extends,
+      c("generic", "generic", "", "")
+    )
+    expect_equal(out$get_value("minidesc")$generic, c("zap", "zap", "print", ""))
+    expect_equal(
+      out$get_value("minidesc")$class, 
+      c("numeric", "character", "qux", "")
+    )
+  }
+)
+
+test_that(
+  "Multiple @describeIn methods and others are combined into a class constructor",
+  {
+    example <- test_path("roxygen-block-describe-in-method-in-const.R")
+    out <- roc_proc_text(rd_roclet(), brio::read_file(example))[[1]]
+
+    expect_equal(out$get_value("minidesc")$extends, c("class", "class", "", ""))
+
+    # more complicated with disambiguated class name "pkg_class"
+    out2 <- roc_proc_text(rd_roclet(), brio::read_file(example))[[2]]
+    expect_equal(out2$get_value("minidesc")$extends, c("class", ""))
+  }
+)
+
+
 test_that("function names are escaped", {
   out <- roc_proc_text(rd_roclet(), "
     #' foo
@@ -108,59 +150,3 @@ test_that("function names are escaped", {
   out
   expect_match(out$get_rd("minidesc"), "\\\\%foo\\\\%")
 })
-
-test_that("minidescription is correctly formatted", {
-  x <- rd_section_minidesc(
-    df = subsection(label = c("foo", "bar"), desc = c("baz", "qux")),
-    method = TRUE,
-    generic = "corge"
-  )
-  expect_equal_strings(
-    format(x),
-    "\\section{Related Functions and Methods}{\n
-    \\subsection{Methods extending \\code{corge} generic (by class):}{
-    \n\\itemize{\n\\item \\code{foo}: baz\n\\item \\code{bar}: qux\n}\n
-    }\n}\n"
-  )
-})
-
-test_that("Multiple @describeIn functions combined into one", {
-  example <- test_path("roxygen-block-describe-in-functions.R")
-  out <- roc_proc_text(rd_roclet(), brio::read_file(example))[[1]]
-
-  expect_equal(out$get_value("minidesc")$method, c(FALSE, FALSE))
-  expect_equal(out$get_value("minidesc")$generic, c("", ""))
-  expect_equal(out$get_value("minidesc")$label, c("square", "cube"))
-})
-
-test_that(
-  "Multiple @describeIn methods and others are combined into a generic", {
-    example <- test_path("roxygen-block-describe-in-method-in-gen.R")
-    out <- roc_proc_text(rd_roclet(), brio::read_file(example))[[1]]
-    expect_equal(out$get_value("minidesc")$method, c(TRUE, TRUE, FALSE, FALSE))
-    expect_equal(out$get_value("minidesc")$generic, c("zap", "zap", "", ""))
-    expect_equal(
-      out$get_value("minidesc")$label,
-      c("numeric", "character", "print.qux", "zap_helper")
-    )
-  }
-)
-
-test_that(
-  "Multiple @describeIn methods and others are combined into a class constructor",
-  {
-    example <- test_path("roxygen-block-describe-in-method-in-const.R")
-    out <- roc_proc_text(rd_roclet(), brio::read_file(example))[[1]]
-
-    expect_equal(out$get_value("minidesc")$method, c(TRUE, TRUE, FALSE, FALSE))
-    expect_equal(out$get_value("minidesc")$generic, rep("", 4))
-    expect_equal(
-      out$get_value("minidesc")$label, 
-      c("print", "format", "format.bar", "is_foo")
-    )
-
-    # more complicated with disambiguated class name "pkg_class"
-    out2 <- roc_proc_text(rd_roclet(), brio::read_file(example))[[2]]
-    expect_equal(out2$get_value("minidesc")$method, c(TRUE, FALSE))
-  }
-)