+++ NEWS.md
@@ -1,5 +1,9 @@
 # roxygen2 (development version)
 
+* The new `@exmaplesIf` tag can be used to create conditional
+  examples. These examples only run if a specified condition
+  holds (#962).
+
 # roxygen2 7.1.1
 
 * When processing cross package markdown links (e.g. `[pkg::fun()]`),