+++ NEWS.md
@@ -4,6 +4,8 @@
   examples. These examples only run if a specified condition
   holds (#962).
 
+* roxygen2 is now licensed as MIT (#1163).
+
 # roxygen2 7.1.1
 
 * When processing cross package markdown links (e.g. `[pkg::fun()]`),